+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2022-04-27 - Edu
+### Added
+- Added libft's **get_array_size**
+### Fixed
+- Fix leaks parser with array size
+
 ## 2022-04-27 - Jorge
 - Norminette passed.
 - BLACK_HOLE added.